Revert "Mandy6"
