+++ lib/redisUtils/cacheOps.js
@@ -1,11 +1,10 @@
-//C:\CPRG306\CapstoneV2\lib\redisUtils\cacheOps.js
+//C:\CPRG306\CapstoneV2\lib\redis.js
 // Import the pre-configured Redis client from another module
 const client = require('../redis');  
 
 // Define an asynchronous function to set a cache value with an optional time-to-live (TTL)
 async function setCache(key, data, ttl = 3600) {
   // Convert the data to a JSON string and store it in Redis with an expiration time
-  //why json file?The JSON format is a lightweight data exchange format that is ideal for storing in redis.
   await client.set(key, JSON.stringify(data), 'EX', ttl);
 }
