+++ lib/redisUtils/index.js
@@ -1,33 +1,10 @@
-// C:\CPRG306\CapstoneV2\lib\redisUtils\index.js 
-// import the setCache and getCache functions from cacheOps.js
+// index.js
 const { setCache, getCache } = require('./cacheOps');
+const { addToList, removeFromList } = require('./listOps');  // 假设你也有listOps.js文件
 
-// Import session-related operations from userps.js
-const { 
-    saveSession, 
-    getSession, 
-    updateSession, 
-    deleteSession, 
-    incrementLoginAttempts 
-} = require('./userOps');
-
-// Import product-related operations from productOps.js
-const {
-    cacheProductInfo,
-    getCachedProductInfo,
-    updateProductCache
-} = require('./productOps');
-
-// Export all imported functions
 module.exports = {
     setCache,
     getCache,
-    saveSession,
-    getSession,
-    updateSession,
-    deleteSession,
-    incrementLoginAttempts,
-    cacheProductInfo,
-    getCachedProductInfo,
-    updateProductCache
+    addToList,
+    removeFromList
 };
