+++ package.json
@@ -18,7 +18,6 @@
     "express-session": "^1.18.0",
     "fs": "^0.0.1-security",
     "knex": "^3.1.0",
-    "multer": "^1.4.5-lts.1",
     "next": "latest",
     "nodemon": "^3.1.4",
     "passport": "^0.7.0",
@@ -28,8 +27,8 @@
     "react": "latest",
     "react-dom": "latest",
     "react-icons": "^5.3.0",
-    "react-query": "^3.39.3",
     "react-slick": "^0.30.2",
+    "react-query": "^3.39.3",
     "redis": "^4.7.0",
     "sequelize": "^6.37.3",
     "slick-carousel": "^1.8.1",
