+++ server/models/Category.js
@@ -1,4 +1,3 @@
-//C:\CPRG306\CapstoneV2\server\models\Category.js
 module.exports = (sequelize, DataTypes) => {
   const Category = sequelize.define('Category', {
     id: {
